Broke out the seach functionality to its own component and created a custom hook for combined data
